added gpt image with
adding textarea and deleting the unnecessary routes
