--- src/routes/home/+page.svelte
@@ -1,3 +1,0 @@
-<h1>main page but not sure what to put it</h1>
-
-
